Add files via upload
fixed logic that had to do with eshkolot in python code file.
Currently the model doesn't produce results, but it seems like it's more of a problem with the mathematical model rather than with the program. further tests is required.
In order to manifest any results, right now I commented the constraint that concerns the eshkolot in the .mod file, and so now when If I'll run the model he will ignore it.