Officially adding the Gini coefficient constraint
After some tests and modification of the prints in the .mod file, it seems that the Gini coefficietns were added successfully to the model.
The mathematical model word and pdf file were updated to include the new constraints and relevant explanations about them.
The python code were also updated and organized. The measurement of running time of the model might still need some tests and improvements.