Modify user interface to change mod file
Adding the editing_mod_file.py module that takes the choices the user makes regarding the model constraints in the interface (i.e. which constraints to remove), and applies them in the logic of the actual running of the opl model, by editing the .mod file that is used by the Agriplots_solve_opl.py module. It does that by using the template.mod and edited.mod files.
In the user interface, also making it so that the user can't remove the Percentage Change in Revenue constraint,
and also he can't choose to leave both the 3rd and 4th constraints, but can either keep one of them or none of them, meaning the constraints that limit energy production by yeshuv and machoz becomes mutually exclusive.