Add option to switch between binary and continous decision variables
Adding a function that get as input the path of the mod file, and the model type, which can be with binary decision variables or continuous decision variables. based on the type of model, the function changes in relevant line in the .mod file that is responsible for setting the variable type (float+ or boolean).