Make single .mod file to be used in the code
Making it so there is only one .mod file that will be used in the code, Agriplots.mod, and his objective fucntion, constraints and decision variables will be modified based on variable values in the code (or from the UI).
There is the new Agriplots_base_model.mod file that contains the main model that has binary decision variables, single objective function and all of the constraints. at the start of the code that file will be copied and pasted as Agriplots.mod, and then Agriplots.mod will be modified.
Therefore, From now on, all changes to the model should be implemented in the Agriplots_base_model.mod file.
Also, added some results that might be exported to the GIS tool of the project in the future. as a result made some changes in the .mod file.