Add new variables for results in .mod file
Added variables of TotalArea and ChangeInRevenue in the .mod file, as preparation for adding new objective function, so that I could always store the objective result in those variables. Besides that, makes the some of the collection of results I made earlier in the .mod file redundant, so I commented it for now, and I'll remove it later.